Address batched eval review comments
